+++ src/domain/types/AuthConfirmTargetField.ts
@@ -1,0 +1,3 @@
+import {UserModel} from '@steroidsjs/nest-modules/user/models/UserModel';
+
+export type AuthConfirmTargetField = Extract<keyof UserModel, 'email' | 'phone'>;