+++ src/infrastructure/module.ts
@@ -6,8 +6,7 @@ import {FileModule} from '@steroidsjs/nest-modules/file/FileModule';
 import {NotifierModule} from '@steroidsjs/nest-modules/notifier/NotifierModule';
 import {IUserService} from '@steroidsjs/nest-modules/user/services/IUserService';
 import {IFileService} from '@steroidsjs/nest-modules/file/services/IFileService';
-import {INotifierService} from '@steroidsjs/nest-modules/notifier/services/INotifierService';
-import {forwardRef} from '@nestjs/common';
+import {forwardRef, ModuleMetadata} from '@nestjs/common';
 import {IAuthUpdateUserOwnPasswordUseCase} from '@steroidsjs/nest-modules/auth/usecases/IAuthUpdateUserOwnPasswordUseCase';
 import {IUserUpdatePasswordUseCase} from '@steroidsjs/nest-modules/user/usecases/IUserUpdatePasswordUseCase';
 import {IAuthRevokeUserActiveLoginsUseCase} from '@steroidsjs/nest-modules/auth/usecases/IAuthRevokeUserActiveLoginsUseCase';
@@ -22,12 +21,17 @@ import {AuthConfirmService} from '../domain/services/AuthConfirmService';
 import {IAuthRoleRepository} from '../domain/interfaces/IAuthRoleRepository';
 import {AuthRoleService} from '../domain/services/AuthRoleService';
 import {AuthFilePermissionService} from '../domain/services/AuthFilePermissionService';
-import {AuthUpdateUserOwnPasswordUseCase} from '../usecases/updatePassword/AuthUpdateUserOwnPasswordUseCase';
-import {AuthRevokeUserActiveLoginsUseCase} from '../usecases/revokeUserActiveLogins/AuthRevokeUserActiveLoginsUseCase';
 import {AuthenticateWithCodeUseCase} from '../usecases/authenticateWithCodeUseCase/AuthenticateWithCodeUseCase';
 import {SendAuthenticationCodeUseCase} from '../usecases/sendAuthenticationCodeUseCase/SendAuthenticationCodeUseCase';
 import {AUTHENTICATE_WITH_CODE_USE_CASE_TOKEN} from '../usecases/authenticateWithCodeUseCase/IAuthenticateWithCodeUseCase';
 import {SEND_AUTHENTICATION_CODE_USE_CASE_TOKEN} from '../usecases/sendAuthenticationCodeUseCase/ISendAuthenticationCodeUseCase';
+import {AuthUpdateUserOwnPasswordUseCase} from '../usecases/updatePassword/AuthUpdateUserOwnPasswordUseCase';
+import {AuthRevokeUserActiveLoginsUseCase} from '../usecases/revokeUserActiveLogins/AuthRevokeUserActiveLoginsUseCase';
+import {AUTH_CONFIRM_PROVIDERS_TOKEN, IAuthConfirmProvider} from '../domain/interfaces/IAuthConfirmProvider';
+import {
+    AUTH_CONFIRM_TARGET_VALIDATORS_TOKEN,
+    IAuthConfirmTargetValidator,
+} from '../domain/interfaces/IAuthConfirmTargetValidator';
 import {SessionService} from './services/SessionService';
 import {AuthLoginRepository} from './repositories/AuthLoginRepository';
 import {AuthPermissionRepository} from './repositories/AuthPermissionRepository';
@@ -36,6 +40,8 @@ import {JwtStrategy} from './strategies/JwtStrategy';
 import {AuthConfirmRepository} from './repositories/AuthConfirmRepository';
 import {LoginSmsCodeStrategy} from './strategies/LoginSmsCodeStrategy';
 import {AuthRoleRepository} from './repositories/AuthRoleRepository';
+import {authConfirmProviders} from './services/authConfirmProviders';
+import {authConfirmTargetValidators} from './services/authConfirmTargetValidators';
 import {AuthController} from './controllers/AuthController';
 import {AuthFilePermissionController} from './controllers/AuthFilePermissionController';
 import {AuthPermissionController} from './controllers/AuthPermissionController';
@@ -44,7 +50,7 @@ import {AuthRoleController} from './controllers/AuthRoleController';
 import {IAuthModuleConfig} from './config';
 import {PasswordValidator} from './validators/PasswordValidator';
 
-export default (config: IAuthModuleConfig) => ({
+export default (config: IAuthModuleConfig): ModuleMetadata => ({
     imports: [
         PassportModule,
         NotifierModule,
@@ -92,11 +98,19 @@ export default (config: IAuthModuleConfig) => ({
             AuthLoginService,
             AuthPermissionsService,
         ]),
-        ModuleHelper.provide(AuthConfirmService, [
-            IAuthConfirmRepository,
-            INotifierService,
-            AuthService,
-        ]),
+        ...authConfirmProviders,
+        ...authConfirmTargetValidators,
+        {
+            provide: AUTH_CONFIRM_PROVIDERS_TOKEN,
+            useFactory: (...providers: IAuthConfirmProvider[]) => providers,
+            inject: authConfirmProviders,
+        },
+        {
+            provide: AUTH_CONFIRM_TARGET_VALIDATORS_TOKEN,
+            useFactory: (...validators: IAuthConfirmTargetValidator[]) => validators,
+            inject: authConfirmTargetValidators,
+        },
+        AuthConfirmService,
         ModuleHelper.provide(AuthLoginService, [
             IAuthLoginRepository,
             ISessionService,