+++ src/infrastructure/services/authConfirmProviders/BaseAuthConfirmProvider.ts
@@ -1,0 +1,31 @@
+import {INotifierService} from '@steroidsjs/nest-modules/notifier/services/INotifierService';
+import {INotifierSendOptions} from '@steroidsjs/nest-modules/notifier/interfaces/INotifierSendOptions';
+import NotifierSendException from '@steroidsjs/nest-modules/notifier/exceptions/NotifierSendException';
+import {ValidationException} from '@steroidsjs/nest/usecases/exceptions/ValidationException';
+import {IAuthConfirmConfig} from 'src/infrastructure/config';
+import {IAuthConfirmProvider} from '../../../domain/interfaces/IAuthConfirmProvider';
+import {AuthConfirmProviderType} from '../../../domain/types/AuthConfirmProviderType';
+
+export abstract class BaseAuthConfirmProvider implements IAuthConfirmProvider {
+    protected constructor(
+        protected readonly notifierService: INotifierService,
+    ) {
+    }
+
+    abstract readonly type: AuthConfirmProviderType;
+
+    abstract generateAndSendCode(config: IAuthConfirmConfig, target: string): Promise<string>;
+
+    protected async sendCode(options: Partial<INotifierSendOptions>) {
+        try {
+            return await this.notifierService.send(options);
+        } catch (e) {
+            if (e instanceof NotifierSendException) {
+                throw new ValidationException({
+                    target: 'Не удалось отправить код',
+                });
+            }
+            throw e;
+        }
+    }
+}