+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ### Fixes
 - Замена ts-типа `Date` на `string` у полей с `DateTimeField` или `StringField` декоратором
+- Отправка кода по разным каналам была вынесена из `AuthConfirmService` в `authConfirmProviders`, провайдятся токеном `AUTH_CONFIRM_PROVIDERS_TOKEN`
+- Функция `generateCode` вынесена из `domain/services/AuthConfirmService.ts` в `domain/utils/index.ts`
+- Добавлен `GetAuthConfirmTargetFieldUseCase` для получения поля из пользователя, которое надо взять для отправки кода подтверждения, провайдится по токену `GET_AUTH_CONFIRM_TARGET_FIELD_USE_CASE_TOKEN`
 
 ## [0.3.0](https://github.com/steroids/nest-auth/compare/0.2.3...0.3.0) (2025-12-26)